Changed the type from int to long. Overshot the int limit. Added tests.

+++ src/main/java/com/sampleproject/amountconverter/service/impl/Converter.java
@@ -24,7 +24,7 @@ public class Converter implements IConverter {
 		if (number > LIMIT)
 			return "WOW!! thats a lot of money. Unfortunately the conversion cannot handle such large amount.";
 
-		int wholeNumber = (int) number;
+		long wholeNumber = (long) number;
 		double fraction = number - wholeNumber;
 
 		if (wholeNumber > 999)
@@ -43,7 +43,7 @@ public class Converter implements IConverter {
 		return sb.toString();
 	}
 
-	private String convertLargeNums(int wholeNumber) {
+	private String convertLargeNums(long wholeNumber) {
 		StringBuffer sb = new StringBuffer();
 		int i = 0;
 		while (wholeNumber > 0) {
@@ -58,16 +58,16 @@ public class Converter implements IConverter {
 		return sb.toString();
 	}
 
-	private String convertTripleDigit(int wholeNumber) {
+	private String convertTripleDigit(long wholeNumber) {
 		StringBuffer sb = new StringBuffer();
-		sb.append(UNITS[wholeNumber / 100]);
+		sb.append(UNITS[(int) (wholeNumber / 100)]);
 		sb.append(" hundred");
-		int num = wholeNumber % 100;
+		long num = wholeNumber % 100;
 		sb.append(convertDoubleDigit(num));
 		return sb.toString();
 	}
 
-	private String convert(int wholeNumber) {
+	private String convert(long wholeNumber) {
 		StringBuffer sb = new StringBuffer();
 		if (wholeNumber == 0)
 			sb.append("Zero");
@@ -81,16 +81,16 @@ public class Converter implements IConverter {
 		return sb.toString();
 	}
 
-	private String convertSingleDigit(int num) {
-		return UNITS[num];
+	private String convertSingleDigit(long num) {
+		return UNITS[(int) num];
 	}
 
-	private String convertDoubleDigit(int num) {
+	private String convertDoubleDigit(long num) {
 		StringBuffer sb = new StringBuffer();
 		if (num < 20)
-			sb.append(UNITS[num]);
+			sb.append(UNITS[(int) num]);
 		else {
-			sb.append(TENS[num / 10]);
+			sb.append(TENS[(int) (num / 10)]);
 			sb.append(convertSingleDigit(num % 10));
 		}
 		return sb.toString();
